transformation problem solved

+++ transformation/transformation.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 
 int main() {
-    int n;
+    long long n;
     cin >> n;
-    int changes = 0;
+    long long changes = 0;
     while (n > 1) {
         if (n % 2 == 0)
         {
